Remove FileMaker hacks
This stops the aggressive icon building for files that are not CSV and should not be processed by this plugin. However it might break FileMaker TSV icon generation (I can no longer test this), if it does please use version 1.2.

+++ CSVDocument.m
@@ -114,7 +114,7 @@
 					isNewColumn = NO;
 				}
 				else {
-					colKey = [NSString stringWithFormat:@"col_%lu", colIndex];
+					colKey = [NSString stringWithFormat:@"col_%lu", (unsigned long)colIndex];
 					isNewColumn = YES;
 				}
